Prueba lseekFile

+++ test.c
@@ -29,7 +29,7 @@ int main()
 {
 	int ret;
 	
-	///////
+	/////// MKFS /////////
 	ret = mkFS(DEV_SIZE);
 	if (ret != 0)
 	{
@@ -38,7 +38,8 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST mkFS ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// MOUNT /////////
 
 	ret = mountFS();
 	if (ret != 0)
@@ -48,7 +49,8 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST mountFS ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// CREATE FILE /////////
 
 	ret = createFile("/test.txt");
 	if (ret != 0)
@@ -58,7 +60,8 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST createFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// CLOSE FILE /////////
 
 	ret = closeFile(0); // El archivo se debe haber creado en el primer inodo libre
 	if (ret != 0)
@@ -68,10 +71,11 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST closeFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// OPEN FILE /////////
 
 	ret = openFile("/test.txt");
-	int descriptor = ret;
+	//int descriptor = ret;
 	if (ret < 0) // Debe ser 0 porque es donde tenemos ese archivo concreto, en el primer inodo que hay 
 	{	
 		printf("Hola: %i", ret);
@@ -80,7 +84,8 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST openFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+/*
+	/////// WRITE FILE /////////
 		
 	char *buffer1 [] = {"Hola, esto es una prueba para el writeFile"};
 	ret = writeFile(descriptor, buffer1, sizeof(buffer1));
@@ -94,7 +99,8 @@ int main()
 	printf("El contenido del buffer1 es: %s\n", *buffer1);
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST writeFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// READ FILE /////////
 
 	char *buffer2 [BLOCK_SIZE];
 	ret = readFile(descriptor, buffer2, 2048);
@@ -106,8 +112,21 @@ int main()
 	printf("Descriptor con readFile vale: %i\n", descriptor);
 	printf("El contenido del buffer2 es: %s\n", *buffer2);
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST readFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
+*/
 
-	///////
+	/////// LSEEK FILE /////////
+
+	ret = lseekFile(0, 10, 0); // Debería llevar el puntero a la posición 0 y devolver un 0
+	if (ret != 0)
+	{
+		fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST removeFile ", ANSI_COLOR_RED, "FAILED\n", ANSI_COLOR_RESET);
+		return -1;
+	}
+	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST removeFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
+	printf("El retorno es: %i\n", ret);
+
+
+	/////// REMOVE FILE /////////
 
 	ret = removeFile("/test.txt");
 	if (ret != 0)
@@ -117,7 +136,8 @@ int main()
 	}
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST removeFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
-	///////
+
+	/////// CLOSE FILE /////////
 
 	ret = closeFile(0); // El archivo  debe ser cerrado antes de hacer unmount
 	if (ret != 0)
@@ -128,6 +148,7 @@ int main()
 	fprintf(stdout, "%s%s%s%s%s", ANSI_COLOR_BLUE, "TEST closeFile ", ANSI_COLOR_GREEN, "SUCCESS\n", ANSI_COLOR_RESET);
 
 
+	////// UNMOUNT /////////
 	ret = unmountFS();
 	if (ret != 0)
 	{
